Handle multiple explicit mappings to the same schema
The previous implementation would return an incorrectly-computed set of
implied mappings for a oneOf schema in the case where there were multiple
explicitly mappings to the same schema ref, since it assumed that any given
schema ref was targeted by exactly one mapping.

This commit changes the mapping computation such that explicitly-provided
mappings are considered first and always added to the result map, and implied
mappings are then added for any schemas that are not the target of an explicit
mapping.

For example, if the schema looks like:

  oneOf:
  - $ref: '#/components/schemas/Cat'
  - $ref: '#/components/schemas/Dog'
  - $ref: '#/components/schemas/Lizard'
  discriminator:
    propertyName: pet_type
    mapping:
      cat: '#/components/schemas/Cat'
      lion: '#/components/schemas/Cat'

The resulting computation will now have mappings for cat, lion, Dog, and Lizard.